refactor(web): tidy submission log table
